backend/requirements.txt
asgiref==3.9.2
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
python-dotenv==1.0.*

# Database drivers
psycopg2-binary==2.9.*

# WebSocket and Real-time Features
channels==4.1.*
channels_redis==4.2.*
daphne==4.1.*

# Redis for caching and message broker
redis==5.2.*

# Payment Processing
stripe==11.3.*
